+++ app/src/main/java/com/example/hundredplaces/data/model/image/repositories/ImagesDataRepository.kt
@@ -17,7 +17,7 @@ class ImagesDataRepository(
         }
     }
 
-    override suspend fun getImageStream(id: Int): Flow<Image?> {
+    override suspend fun getImageStream(id: Long): Flow<Image?> {
         return if (networkConnection.isNetworkConnected) {
             imagesRemoteRepository.getImageStream(id)
         } else {
